Refactor player ranking updater into reusable class
